+++ rgrow/src/models/atam.rs
@@ -12,6 +12,7 @@ use pyo3::prelude::*;
 use crate::base::{HashMapType, HashSetType};
 use ndarray::prelude::*;
 use rand::prelude::Distribution;
+use rand::distr::Uniform;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
@@ -683,8 +684,8 @@ impl ATAM {
         atam.tile_colors = match tile_colors {
             Some(tc) => tc,
             None => {
-                let mut rng = rand::thread_rng();
-                let ug = rand::distributions::Uniform::new(100u8, 254);
+                let mut rng = rand::rng();
+                let ug = rand::distr::Uniform::new(100u8, 254).unwrap();
                 (0..ntiles)
                     .map(|_x| {
                         [